Setup base application
Might be my lack of experience, but boy is it more complex to set up
a TypeScript app from scratch compared to something in plain JS or Go.

Fought (and I think I won?) with the tsconfig.json file, and then
struggled with module resolution for a graceful shutdown lib which seems
to use the wrong module types.

This reminded me of the manual webpack setup from way-back-when.