Setup models and index pokemon on startup
While the indexing isn't perfect, the approach taken here at least only
updates what's new. Ideally, I would like to persist more often so that
the initial indexing _could_ be interrupted and still persist partial
data instead of this all-or-nothing approach. For later, though!

I am again finding myself fighting TS more than I would like. Kinda feel
like I'm creating a bit of spaghetti with the type imports compared to
what I'm used to with other languages (or not having in JS or Python 🙃)

To not duplicate too much of the typing, I'm opting for a global set of
types that are shared across the application that kinda serves as the
_domain types_, they also happen to be auto generated types from the
pokeapi, which ideally would have their own types that we'd map from
the pokeapi types to our own domain types. But again, that's for later.